backend/requirements.txt
fastapi==0.129.1
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
pydantic[email]==2.12.0
python-dotenv==1.0.1
openai==1.85.0
# CrewAI/LangChain desabilitado por conflito - ativamos depois
# crewai==0.67.1
# langchain==0.2.16
# langchain-community==0.0.38
pypdf==5.6.0
slowapi==0.1.9
alembic==1.16.2
bcrypt==4.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
